pyproject.toml
[tool.poetry]
name = "test-task-for-cft"
version = "0.1.0"
description = ""
authors = ["DeadDarvin"]
readme = "README.md"
packages = [{include = "test_task_for_cft"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
SQLAlchemy = "^1.4.45"
pydantic = {version = "*", extras = ["email"]}
envparse = "^0.2.0"
psycopg2-binary = "^2.9.5"
asyncpg = "^0.27.0"
alembic = "^1.9.0"
pytest = "^7.2.0"
pytest-asyncio = "^0.20.3"
httpx = "^0.23.3"
pre-commit = "^2.21.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.5"
bcrypt = "^4.0.1"
greenlet = "^2.0.2"
sentry-sdk = {version = "*", extras = ["fastapi"]}
starlette-exporter = "^0.15.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
